% Auto-generated: do not edit by hand
\name{RCircularGenomeView}

\alias{RCircularGenomeView}

\title{CircularGenomeView component}

\description{
CircularGenomeView renders the JBrowse 2 React Circular Genome View. Any JB2 CGV configuration is also valid configuration for this component.
}

\usage{
RCircularGenomeView(id=NULL, aggregateTextSearchAdapters=NULL, assembly=NULL,
configuration=NULL, createRootFn=NULL, defaultSession=NULL,
hydrateFn=NULL, internetAccounts=NULL,
makeWorkerInstance=NULL, onChange=NULL, plugins=NULL,
tracks=NULL)
}

\arguments{
\item{id}{Character. }

\item{aggregateTextSearchAdapters}{List of logical | numeric | character | named list | unnamed lists. }

\item{assembly}{Logical | numeric | character | named list | unnamed list. }

\item{configuration}{Logical | numeric | character | named list | unnamed list. }

\item{createRootFn}{}

\item{defaultSession}{Logical | numeric | character | named list | unnamed list. }

\item{hydrateFn}{}

\item{internetAccounts}{List of logical | numeric | character | named list | unnamed lists. }

\item{makeWorkerInstance}{}

\item{onChange}{}

\item{plugins}{Logical | numeric | character | named list | unnamed list. }

\item{tracks}{List of logical | numeric | character | named list | unnamed lists. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

