% Auto-generated: do not edit by hand
\name{RLinearGenomeView}

\alias{RLinearGenomeView}

\title{LinearGenomeView component}

\description{

}

\usage{
RLinearGenomeView(id=NULL, aggregateTextSearchAdapters=NULL, assembly=NULL,
configuration=NULL, createRootFn=NULL, defaultSession=NULL,
disableAddTracks=NULL, hydrateFn=NULL,
internetAccounts=NULL, location=NULL,
makeWorkerInstance=NULL, onChange=NULL, plugins=NULL,
tracks=NULL)
}

\arguments{
\item{id}{Character. }

\item{aggregateTextSearchAdapters}{List of logical | numeric | character | named list | unnamed lists. }

\item{assembly}{Logical | numeric | character | named list | unnamed list. }

\item{configuration}{Logical | numeric | character | named list | unnamed list. }

\item{createRootFn}{}

\item{defaultSession}{Logical | numeric | character | named list | unnamed list. }

\item{disableAddTracks}{Logical | numeric | character | named list | unnamed list. }

\item{hydrateFn}{}

\item{internetAccounts}{List of logical | numeric | character | named list | unnamed lists. }

\item{location}{Character. }

\item{makeWorkerInstance}{}

\item{onChange}{}

\item{plugins}{Logical | numeric | character | named list | unnamed list. }

\item{tracks}{List of logical | numeric | character | named list | unnamed lists. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

