% Auto-generated: do not edit by hand
\name{dashJbrowse}

\alias{dashJbrowse}

\title{DashJbrowse component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashJbrowse(id=NULL, assembly=NULL, defaultSession=NULL, location=NULL,
tracks=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{assembly}{Named list. The configuration for the assembly to use for the browser.}

\item{defaultSession}{Named list. The default session to use for the browser.}

\item{location}{Character. The location to use for the browser.}

\item{tracks}{Unnamed list. The tracks to use for the browser.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

